requirements.txt
# use_rag dependencies
#
# Level 1-3: No required dependencies (pure Python)
#
# Level 2 (optional): Config file loading
pyyaml>=6.0

# Level 4 (optional): LLM client - install ONE of these
litellm>=1.0    # Supports multiple providers (OpenAI, Anthropic, etc.)
# openai>=1.0     # OpenAI only

# Level 4 (optional): Environment file support
python-dotenv>=1.0
